% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronologia.R
\name{create_timeline}
\alias{create_timeline}
\title{Create a Vertical Time Line}
\usage{
create_timeline(
  df,
  smr,
  dsc,
  smr_col = "white",
  smr_bgcol = "#002240",
  dsc_col = "#1D1D1D",
  dsc_bgcol = "#FFFFFF"
)
}
\arguments{
\item{df}{a data frame containing the necessary information}

\item{smr}{the column that will be used as the summary component}

\item{dsc}{the column that will be used as the description component}

\item{smr_col}{text color for the summary component. Defaults to 'white'}

\item{smr_bgcol}{background-color for the summary component. Defaults to "#002240"}

\item{dsc_col}{text color for the description component. Defaults to "#1D1D1D"}

\item{dsc_bgcol}{background-color for the description component. Defaults to "#FFFFFF"}
}
\value{

}
\description{
Create a Vertical Time Line
}
