% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronologia.R
\name{create_tml_2}
\alias{create_tml_2}
\title{Create a Vertical HTML Timeline with 2 Description Levels from a Data Frame}
\usage{
create_tml_2(
  df,
  smr,
  dsc,
  dsc2,
  smr_col = "white",
  smr_bgcol = "#002240",
  dsc_col = "#1D1D1D",
  dsc_bgcol = "#FFFFFF",
  dsc_size = "14px",
  dsc_col2 = "#007FF1",
  dsc_bgcol2 = "#F45E88",
  dsc_size2 = "10px"
)
}
\arguments{
\item{df}{a data frame containing the necessary information.}

\item{smr}{the column name that will be used as the summary component.}

\item{dsc}{the column name that will be used as the first description component.}

\item{dsc2}{the column name that will be used as the second description component.}

\item{smr_col}{text color for the summary component. Defaults to 'white'.}

\item{smr_bgcol}{background-color for the summary component. Defaults to "#002240".}

\item{dsc_col}{text color for the first description component. Defaults to "#1D1D1D".}

\item{dsc_bgcol}{background-color for the first description component. Defaults to "#FFFFFF".}

\item{dsc_size}{font size of the first description component. Defaults to "14px".}

\item{dsc_col2}{text color for the second description component. Defaults to "#007FF1".}

\item{dsc_bgcol2}{background-color for the second description component. Defaults to "#F45E88".}

\item{dsc_size2}{font size of the second description component. Defaults to "10px".}
}
\value{
A Vertical HTML Time Line Widget with 2 Description Levels
}
\description{
Create a Vertical HTML Timeline with 2 Description Levels from a Data Frame
}
