% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronologia.R
\name{create_tml_img}
\alias{create_tml_img}
\title{Title}
\usage{
create_tml_img(
  df,
  smr,
  dsc,
  imgsrc,
  imgalt,
  smr_col = "white",
  smr_bgcol = "#002240",
  dsc_col = "#1D1D1D",
  dsc_bgcol = "#FFFFFF",
  imgheight = "150px",
  imgwidth = "150px"
)
}
\arguments{
\item{imgwidth}{}
}
\value{

}
\description{
Title
}
