% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronologia.R
\name{create_tml_img}
\alias{create_tml_img}
\title{Create Vertical HTML Timeline Widget with Images from a Data Frame}
\usage{
create_tml_img(
  df,
  smr,
  dsc,
  imgsrc,
  imgalt,
  smr_col = "white",
  smr_bgcol = "#002240",
  dsc_col = "#1D1D1D",
  dsc_bgcol = "#FFFFFF",
  dsc_size = "14px",
  imgheight = "150px",
  imgwidth = "150px",
  open = FALSE
)
}
\arguments{
\item{df}{a data frame containing the necessary information.}

\item{smr}{the column name that will be used as the summary component.}

\item{dsc}{the column name that will be used as the description component.}

\item{imgsrc}{the column name indicating the 'href' source of the images.}

\item{imgalt}{the column name that will be used as 'alt' text for the images.}

\item{smr_col}{text color for the summary component. Defaults to 'white'.}

\item{smr_bgcol}{background-color for the summary component. Defaults to "#002240".}

\item{dsc_col}{text color for the description component. Defaults to "#1D1D1D".}

\item{dsc_bgcol}{background-color for the description component. Defaults to "#FFFFFF".}

\item{dsc_size}{font size of the description component. Defaults to "14px".}

\item{imgheight}{the height of the image. Defaults to "150px".}

\item{imgwidth}{the width of the image. Defaults to "150px".}

\item{open}{logical, whether to open by default the summary components. Default to FALSE}
}
\value{
A Vertical HTML Time Line Widget With Images
}
\description{
Create Vertical HTML Timeline Widget with Images from a Data Frame
}
